Opening moves - runtime config read
